Use username instead possibly null display name
Display names don't exist for users who have left the server, which
causes a null exception trying to pin their messages. This commit fixes
that by using the author username instead of the displayname. This means
we loses out on the display name, but it didn't make sense to use
displayname#discriminator anyway.